WIP on event-specific filters.
Checking in WIP towards event specific filters.

Instead of combining all rules into one huge filter expression and
giving it to the inspector, keep each filter expression separate and
annotate it with the events for which the rule applies.

Within the compiler, a new pass over the ast get_evttypes looks for
evt.type clauses, converts the evt.type as a string to any event type
ids for which it may apply, and passes that back with the compiled rule.

In the rule loader, create a new global events that maps each event as a
string to the list of event ids for which it may apply. Also there's a
callback add_filter from lua which is called for each rule. In turn, it
passes the rule and list of event ids to the inspector using
add_evttype_filter().

This implies a new restriction, that each rule must name an event type
for which the rule applies.

TODO:

 - There are some remnants of the old parsing code left in the rule
 - loader that should be removed.
 - The string splitting within the compiler is ugly and should be
   cleaned up.
 - not evt.type=XXX isn't handled right now.
 - or clauses combined with event conditions (evt.type=XXX or
   evt.args=XXX) are mistakenly restricted to the event condition.
 - Should we just prevent use of syscall.type entirely? If so, we don't
   really need the ignored syscall table.